backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
elevenlabs==0.2.27
pydantic==2.5.0
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
httpx==0.25.1
